Base things tested

+++ src/main/java/com/example/vendingmachine/controller/UserController.java
@@ -2,21 +2,28 @@ package com.example.vendingmachine.controller;
 
 import com.example.vendingmachine.model.SignUpData;
 import com.example.vendingmachine.model.entity.User;
+import com.example.vendingmachine.service.SecurityService;
 import com.example.vendingmachine.service.UserService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import javax.validation.Valid;
 import java.util.List;
+import java.util.stream.IntStream;
+import java.util.stream.Stream;
 
 ;
 
 @RestController
 @RequestMapping("/api/user")
 public class UserController {
+
+    @Autowired
+    private SecurityService securityService;
 
     @Autowired
     private UserService userService;
@@ -29,9 +36,13 @@ public class UserController {
      */
 
     @PostMapping
-    public ResponseEntity<?> signUp(@Valid SignUpData signUpData) {
+    public ResponseEntity<?> signUp(@Valid @RequestBody SignUpData signUpData) {
         User user = new User();
         BeanUtils.copyProperties(signUpData, user);
+
+        if(!Stream.of(User.POSSIBLE_ROLES).anyMatch(x -> x.equals(user.getRole())))
+            return new ResponseEntity<Object>("Wrong user role.", HttpStatus.NOT_ACCEPTABLE);
+
         userService.saveUser(user);
         return ResponseEntity.ok(user);
     }
@@ -87,9 +98,22 @@ public class UserController {
      * @return
      */
 
-    @PreAuthorize("authentication.id == #userId")
+    @PreAuthorize("isAuthenticated()")
     @PutMapping(value = "/{id}")
-    public ResponseEntity<?> updateUser(@PathVariable("id") Long userId, @RequestBody User user) {
+    public ResponseEntity<?> updateUser(@PathVariable("id") Long userId, @RequestBody SignUpData signUpData) {
+        //todo solve - cannot authorize after changing name or password
+        User loggedInUser = securityService.findLoggedInUser();
+        if (loggedInUser == null)
+            return new ResponseEntity<Object>("User does not exist.", HttpStatus.UNAUTHORIZED);
+        if (loggedInUser.getId() != userId)
+            return new ResponseEntity<Object>("Please use your user ID.", HttpStatus.UNAUTHORIZED);
+
+        User testUser = userService.findByUsername(signUpData.getUsername());
+        if(testUser.getId() != loggedInUser.getId())
+            return new ResponseEntity<Object>("Username already used.", HttpStatus.NOT_ACCEPTABLE);
+
+        User user = new User();
+        BeanUtils.copyProperties(signUpData, user);
         user.setId(userId);
         User updatedUser = userService.updateUser(user);
         return ResponseEntity.ok(updatedUser);
@@ -99,10 +123,17 @@ public class UserController {
      * @param userId
      * @return
      */
-    @PreAuthorize("authentication.id == #userId")
+    @PreAuthorize("isAuthenticated()")
     @DeleteMapping(value = "/{id}")
     public ResponseEntity<?> deleteUser(@PathVariable("id") Long userId) {
-        //todo check products for sellers and rest of coins for buyers
+        User loggedInUser = securityService.findLoggedInUser();
+
+        if (loggedInUser == null)
+            return new ResponseEntity<Object>("User does not exist.", HttpStatus.UNAUTHORIZED);
+        if (loggedInUser.getId() != userId)
+            return new ResponseEntity<Object>("Please use your user ID.", HttpStatus.UNAUTHORIZED);
+
+        //todo check the products if seller is removed
         if(userService.deleteUser(userId)) {
             return ResponseEntity.ok("User '" + userId + "' has been deleted successfully.");
         }
